fixed errors. added validation. removed potential panic from query calls
